Polish transcript navigation architecture
